+++ Plugin/ReplaceTagsPlugin.php
@@ -3,8 +3,8 @@
 namespace Yireo\NextGenImages\Plugin;
 
 use Magento\Framework\View\LayoutInterface;
-use Yireo\NextGenImages\Config\Config;
 use Yireo\NextGenImages\Util\HtmlReplacer;
+use Yireo\NextGenImages\Util\ShouldModifyOutput;
 
 class ReplaceTagsPlugin
 {
@@ -14,22 +14,22 @@ class ReplaceTagsPlugin
     private $htmlReplacer;
 
     /**
-     * @var Config
+     * @var ShouldModifyOutput
      */
-    private $config;
-
+    private $shouldModifyOutput;
+    
     /**
      * ReplaceTags constructor.
      *
      * @param HtmlReplacer $htmlReplacer
-     * @param Config $config
+     * @param ShouldModifyOutput $shouldModifyOutput
      */
     public function __construct(
         HtmlReplacer $htmlReplacer,
-        Config $config
+        ShouldModifyOutput $shouldModifyOutput
     ) {
         $this->htmlReplacer = $htmlReplacer;
-        $this->config = $config;
+        $this->shouldModifyOutput = $shouldModifyOutput;
     }
 
     /**
@@ -41,43 +41,10 @@ class ReplaceTagsPlugin
      */
     public function afterGetOutput(LayoutInterface $layout, string $output): string
     {
-        if (!$this->config->enabled()) {
-            return $output;
-        }
-
-        if ($this->shouldModifyOutput($layout) === false) {
+        if ($this->shouldModifyOutput->shouldModifyOutput($layout) === false) {
             return $output;
         }
 
         return $this->htmlReplacer->replace($output);
-    }
-
-    /**
-     * @param LayoutInterface $layout
-     * @return bool
-     */
-    private function shouldModifyOutput(LayoutInterface $layout): bool
-    {
-        $handles = $layout->getUpdate()->getHandles();
-        if (empty($handles)) {
-            return false;
-        }
-
-        foreach ($handles as $handle) {
-            if (strstr($handle, '_email_')) {
-                return false;
-            }
-        }
-
-        $skippedHandles = [
-            'webp_skip',
-            'nextgenimages_skip',
-        ];
-
-        if (array_intersect($skippedHandles, $handles)) {
-            return false;
-        }
-
-        return true;
     }
 }